backend/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-environ==0.4.5
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
pkg-resources==0.0.0
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
